API: Add mathematical function used to generate stimuli.
Either standard or ARM platform implementation are available as part of
the API so it can be used with non ARM platforms although it is not the
intended  purpose. Also helper method to convert from radians to samples
and to generate the click stimuli as a periodic signal in radians are
available.