Stim: Add implementation for single frequency tones.
It generates the samples using the sine mathematical function.
It also marks the sample corresponding to the 2*PI period of
the signal. The algorithm can generate any amount of samples
on each iteration of the generation process. On the next iteration
it will continue on the next sample because it stores the last index
sample on each generation process.